backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Backend para o projeto DeGym zkVerify"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]>=0.27.1",
    "pydantic==2.6.3",
    "pydantic-settings==2.2.1",
    "python-dotenv==1.0.1",
    "web3==6.11.3",
    "setuptools>=61.0",
    "eth-typing<4.0.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project.optional-dependencies]
dev = [
    "pytest==8.3.5",
    "httpx==0.27.0",
    "black==24.3.0",
    "ruff==0.3.4",
    "pytest-cov==4.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
exclude = ["tests*"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.black]
line-length = 88
target-version = ["py310"]
